running on RedBoard Turbo
faster, more space. The only thing I was missing before was pullups on the I2C lines... Added upload check when idle, so Viewer can request latest files.

On physical side, soldered up a SD card breakout to an Arduino proto sheild, wired to SPI port. I think there's a better way to do this with the assignable IO ports but this works for now

starting to need a setup file....